Basic integration with the download manager is mostly complete; products for a scenario can be downloaded.
After products are downloaded, the default script is run to register
the products in the ODA server.
Includes code to split up a product download into metadata and data.
Part of this functionality may be moved to a DM plugin later.
Other improvements include some feedback via the progress bar.